modify: getRootHash
